Update y Delete Perfil

+++ app.py
@@ -180,15 +180,17 @@ def perfiles():
 
 # Actualizar datos de /Perfiles
 @app.route('/updateperfil/<int:id>', methods=['GET', 'POST'])
+@login_required
 def update_perfiles(id):
+    print("Hello")
     error=None
     users = User.query.all()
     user_to_update = User.query.get_or_404(id)
 
     if request.method == "POST":
+        user_to_update.username = request.form['username']
         user_to_update.name = request.form['name']
         user_to_update.surname = request.form['surname']
-        #email = request.form['email']
         user_to_update.password = request.form['password']
         user_to_update.rol = request.form['rol']
 
@@ -197,21 +199,23 @@ def update_perfiles(id):
             return redirect(url_for('perfiles'))
         except:
             error = 'No se pudo actualizar al usuario.'
-            return render_template("updateperfil.html", error=error, users=users)  
-    else:
-        return render_template("updateperfil.html", error=error, users=users)
+            return render_template("perfiles.html", error=error, users=users)  
+    
+    return render_template("perfiles.html", error=error, users=users)
 
 # Borrar datos de /Perfiles
-@app.route('/deleteperfil/<int:id>')
+@app.route('/deleteperfil/<int:id>', methods=['GET', 'POST'])
+@login_required
 def delete_perfiles(id):
     user_to_delete = User.query.get_or_404(id)
-
-    try:
-        db.session.delete(user_to_delete)
-        db.session.commit()
-        return redirect(url_for('perfiles'))
-    except:
-        return "Hubo un error borrando al usuario."
+    if request.method == "POST":
+        try:
+            db.session.delete(user_to_delete)
+            db.session.commit()
+            flash('Se ha eliminado exitosamente.')
+            return redirect(url_for('perfiles'))
+        except:
+            return "Hubo un error borrando al usuario."
 
 # Datos del Productor (requiere iniciar sesión)
 @app.route('/productor')
